+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Errors returned by this package provide also optional details map accessible
+  through `detailer` interface.
+
 ## [0.3.0] - 2022-01-03
 
 ### Changed